pyproject.toml
    [project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "abdulwasea89", email = "abdulwasea853@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.109.2",
    "uvicorn[standard]==0.34.3",
    "gunicorn==21.2.0",
    "openai==1.90.0",
    "openai-agents==0.0.13",
    "google-cloud-firestore==2.21.0",
    "google-cloud-storage==3.1.1",
    "firebase-admin==6.9.0",
    "python-dotenv==1.1.0",
    "requests==2.32.4",
    "orjson==3.10.18",
    "httpx==0.28.1",
    "urllib3==2.2.1",
    "pydantic==2.10.4",
    "pydantic-settings==2.5.2",
    "tenacity==9.1.2",
    "colorlog==6.9.0",
    "loguru==0.7.2",
    "fuzzywuzzy==0.18.0",
    "rapidfuzz==3.13.0",
    "langdetect==1.0.9",
    "nltk==3.9.1",
    "deep-translator==1.11.4",
    "psutil==6.1.0",
    "pyinstrument>=4.8.0",
    "shopify-api==0.7.1",
    "shopifyapi==12.7.0",
    "redis==6.2.0",
    "asgiref==3.8.1",
    "mangum==0.19.0",
    "zoko==0.0.1",
    "sentence-transformers",
    "faiss-cpu>=1.7.0",
    "pandas==2.2.2",
]

[project.scripts]
automation = "src.main_simple:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
